Move BaseNutrientReport into its subclasses
