+++ custom/inddex/reports/nutrient_stats.py
@@ -1,21 +1,57 @@
 from memoized import memoized
 
+from custom.inddex.filters import (
+    AgeRangeFilter,
+    BreastFeedingFilter,
+    GenderFilter,
+    PregnancyFilter,
+    RecallStatusFilter,
+    SettlementAreaFilter,
+    SupplementsFilter,
+)
 from custom.inddex.ucr.data_providers.summary_statistics_data import (
     SummaryStatsNutrientDataProvider,
 )
-from custom.inddex.utils import BaseNutrientReport
+from custom.inddex.utils import MultiTabularReport
 
 
-class NutrientStatsReport(BaseNutrientReport):
+class NutrientStatsReport(MultiTabularReport):
     title = 'Output 4 - Nutrient Intake Summary Statistics'
     name = title
     slug = 'nutrient_stats'
 
     @property
+    def fields(self):
+        return super().fields + [
+            GenderFilter,
+            AgeRangeFilter,
+            PregnancyFilter,
+            BreastFeedingFilter,
+            SettlementAreaFilter,
+            SupplementsFilter,
+            RecallStatusFilter
+        ]
+
+    @property
     def report_config(self):
         report_config = super().report_config
-        report_config.update(self.filters_config)
+        report_config.update(self._filters_config)
         return report_config
+
+    @property
+    def _filters_config(self):
+        request_slugs = [
+            'gender',
+            'age_range',
+            'pregnant',
+            'breastfeeding',
+            'urban_rural',
+            'supplements',
+            'recall_status',
+        ]
+        filters_config = super().report_config
+        filters_config.update({slug: self.request.GET.get(slug, '') for slug in request_slugs})
+        return filters_config
 
     @property
     @memoized
